+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KGHK7Y8746SAXPEC9M6VNF4H"
+        "01KGHWGRBJP250E00Q0C72C86Q"
     )
     r"""The unique identifier of the data source"""