+++ packages/orq-rc/RELEASES.md
@@ -5219,3 +5219,13 @@ Based on:
 - [python v4.3.0-rc.20] packages/orq-rc
 ### Releases
 - [PyPI v4.3.0-rc.20] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.20 - packages/orq-rc
+
+## 2026-02-03 13:57:18
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.700.2 (2.801.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.3.0-rc.21] packages/orq-rc
+### Releases
+- [PyPI v4.3.0-rc.21] https://pypi.org/project/orq-ai-sdk/4.3.0-rc.21 - packages/orq-rc